Remove base controller, rename base url, remove html static, refactor pipeline
