+++ src/main/kotlin/io/github/anki/anki/AnkiApplication.kt
@@ -2,11 +2,9 @@ package io.github.anki.anki
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
-import org.springframework.web.bind.annotation.RestController
 
 
 @SpringBootApplication
-@RestController("/")
 class AnkiApplication
 
 fun main(args: Array<String>) {
