+++ src/main/kotlin/io/github/anki/anki/models/Card.kt
@@ -1,11 +1,5 @@
 package io.github.anki.anki.models
 
-data class CardsCollection(
-    var id: Int,
-    var name: String,
-    var description: String,
-)
-
 data class Card(
     var id: Int,
     var parentCollectionId: Int,
